refactor: pieces 폴더구조 해체
